[IMP] helpdesk_mgmt: separate filters to force AND logic
When filters are not separated with another element, their domains are combined
with an OR selection. This is almost certainly not what you want. If you select
Open Tickets end My tickets, you expect your tickets that are also open, not a
combination of open tickets and tickets assigned to you.